+++ packages/core/common/utils.ts
@@ -159,26 +159,8 @@ export function hookPrototype<T extends HasPrototype, K extends keyof T['prototy
     proto && hookFn(proto, funcName, hookedFunc);
 }
 
-export function restoreHookedFn<T extends object, K extends keyof T>(target: T, funcName: K) {
-    let fn = target[funcName] as AnyFunction;
-    if (typeof fn !== 'function') return;
-
-    while (Object.hasOwn(fn, 'rawFunction')) {
-        fn = (fn as any).rawFunction;
-    }
-    (target[funcName] as any) = fn;
-}
-
 export const checkEnv = () =>
     typeof window !== 'undefined' && window.sea != undefined && window.sea?.CoreReady && window.sea?.SeerH5Ready;
-
-export const extractObjectId = <T extends { [key in K]: number }, K extends string>(obj: T | number, key: K) => {
-    if (typeof obj === 'number') {
-        return obj;
-    } else {
-        return obj[key];
-    }
-};
 
 export const NOOP = () => {};