+++ packages/core/engine/PlayerData.ts
@@ -1,10 +1,9 @@
-import { extractObjectId } from '../common/utils.js';
+import { EntityBase } from '../entity/EntityBase.js';
 import type { IItemObject, Pet } from '../entity/index.js';
-import { Item } from '../entity/index.js';
 import * as Socket from './socket.js';
 
 export async function getItemNum(item: number | IItemObject) {
-    const id = extractObjectId(item, Item.instanceKey);
+    const id = EntityBase.inferId(item);
     await new Promise((res) => ItemManager.updateItems([id], res));
     return ItemManager.getNumByID(id);
 }