+++ packages/core/entity/Item.ts
@@ -7,8 +7,6 @@ export interface IItemObject {
 }
 
 export class Item extends EntityBase implements IItemObject {
-    static readonly key = 'id';
-    static readonly instanceKey = 'id';
     declare readonly __type: EntityType;
     limit?: number;
     amount?: number;