+++ packages/server/src/router/trpc.ts
@@ -1,5 +1,6 @@
 import { initTRPC } from '@trpc/server';
-import { Context } from './context.ts';
+import superjson from 'superjson';
+import type { Context } from '../context.ts';
 
 const t = initTRPC.context<Context>().create({
     transformer: superjson,
@@ -9,4 +10,4 @@ const t = initTRPC.context<Context>().create({
 });
 
 export const router = t.router;
-export const publicProcedure = t.procedure;
+export const procedure = t.procedure;